monitoring/alerts.py
import asyncio
import aiohttp
import smtplib
from email.mime.text import MimeText
from email.mime.multipart import MimeMultipart
import discord
import telegram
import slack_sdk
from typing import Dict, List, Optional
import json
import logging
from dataclasses import dataclass
from enum import Enum
import time

class AlertSeverity(Enum):
    LOW = "low"
    MEDIUM = "medium"
    HIGH = "high"
    CRITICAL = "critical"

@dataclass
class Alert:
    id: str
    timestamp: float
    severity: AlertSeverity
    component: str
    message: str
    details: Dict
    resolved: bool = False

class AlertManager:
    def __init__(self, config):
        self.config = config
        self.active_alerts = {}
        self.alert_history = []
        self.notification_channels = {}
        
    async def initialize(self):
        await self._setup_notification_channels()
        asyncio.create_task(self._alert_monitoring_loop())
        
    async def _setup_notification_channels(self):
        # Email
        if 'email' in self.config['notifications']:
            self.notification_channels['email'] = EmailNotifier(self.config['notifications']['email'])
            
        # Discord
        if 'discord' in self.config['notifications']:
            self.notification_channels['discord'] = DiscordNotifier(self.config['notifications']['discord'])
            
        # Telegram
        if 'telegram' in self.config['notifications']:
            self.notification_channels['telegram'] = TelegramNotifier(self.config['notifications']['telegram'])
            
        # Slack
        if 'slack' in self.config['notifications']:
            self.notification_channels['slack'] = SlackNotifier(self.config['notifications']['slack'])
            
    async def send_alert(self, severity: AlertSeverity, component: str, message: str, details: Dict = None):
        alert_id = f"{component}_{int(time.time())}"
        
        alert = Alert(
            id=alert_id,
            timestamp=time.time(),
            severity=severity,
            component=component,
            message=message,
            details=details or {}
        )
        
        self.active_alerts[alert_id] = alert
        self.alert_history.append(alert)
        
        # Send notifications based on severity
        if severity == AlertSeverity.CRITICAL:
            await self._send_to_all_channels(alert)
        elif severity == AlertSeverity.HIGH:
            await self._send_to_priority_channels(alert)
        else:
            await self._send_to_standard_channels(alert)
            
    async def _send_to_all_channels(self, alert: Alert):
        for channel_name, notifier in self.notification_channels.items():
            try:
                await notifier.send_alert(alert)
            except Exception as e:
                logging.error(f"Failed to send alert via {channel_name}: {e}")
                
    async def _send_to_priority_channels(self, alert: Alert):
        priority_channels = ['email', 'telegram', 'discord']
        for channel_name in priority_channels:
            if channel_name in self.notification_channels:
                try:
                    await self.notification_channels[channel_name].send_alert(alert)
                except Exception as e:
                    logging.error(f"Failed to send alert via {channel_name}: {e}")
                    
    async def _send_to_standard_channels(self, alert: Alert):
        if 'slack' in self.notification_channels:
            try:
                await self.notification_channels['slack'].send_alert(alert)
            except Exception as e:
                logging.error(f"Failed to send alert via slack: {e}")
                
    async def _alert_monitoring_loop(self):
        while True:
            try:
                await self._check_system_health()
                await self._check_trading_performance()
                await self._check_risk_limits()
                await asyncio.sleep(30)
            except Exception as e:
                logging.error(f"Alert monitoring error: {e}")
                await asyncio.sleep(30)
                
    async def _check_system_health(self):
        # Check CPU, Memory, GPU usage
        # Check database connectivity
        # Check exchange connectivity
        pass
        
    async def _check_trading_performance(self):
        # Check win rate
        # Check PnL trends
        # Check execution times
        pass
        
    async def _check_risk_limits(self):
        # Check VaR limits
        # Check position sizes
        # Check drawdown
        pass

class EmailNotifier:
    def __init__(self, config):
        self.smtp_server = config['smtp_server']
        self.smtp_port = config['smtp_port']
        self.username = config['username']
        self.password = config['password']
        self.recipients = config['recipients']
        
    async def send_alert(self, alert: Alert):
        subject = f"[{alert.severity.value.upper()}] Quantum Trading Alert: {alert.component}"
        
        body = f"""
        Alert Details:
        
        Severity: {alert.severity.value.upper()}
        Component: {alert.component}
        Message: {alert.message}
        Timestamp: {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(alert.timestamp))}
        
        Details:
        {json.dumps(alert.details, indent=2)}
        
        ---
        Quantum Trading System
        """
        
        msg = MimeMultipart()
        msg['From'] = self.username
        msg['To'] = ', '.join(self.recipients)
        msg['Subject'] = subject
        msg.attach(MimeText(body, 'plain'))
        
        try:
            server = smtplib.SMTP(self.smtp_server, self.smtp_port)
            server.starttls()
            server.login(self.username, self.password)
            server.send_message(msg)
            server.quit()
        except Exception as e:
            logging.error(f"Email notification failed: {e}")

class DiscordNotifier:
    def __init__(self, config):
        self.webhook_url = config['webhook_url']
        self.bot_token = config.get('bot_token')
        
    async def send_alert(self, alert: Alert):
        color_map = {
            AlertSeverity.LOW: 0x00ff00,
            AlertSeverity.MEDIUM: 0xffff00,
            AlertSeverity.HIGH: 0xff8800,
            AlertSeverity.CRITICAL: 0xff0000
        }
        
        embed = {
            "title": f"🚨 {alert.severity.value.upper()} Alert",
            "description": alert.message,
            "color": color_map[alert.severity],
            "fields": [
                {"name": "Component", "value": alert.component, "inline": True},
                {"name": "Timestamp", "value": time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(alert.timestamp)), "inline": True}
            ],
            "footer": {"text": "Quantum Trading System"}
        }
        
        if alert.details:
            embed["fields"].append({
                "name": "Details",
                "value": f"```json\n{json.dumps(alert.details, indent=2)[:1000]}```",
                "inline": False
            })
            
        payload = {"embeds": [embed]}
        
        async with aiohttp.ClientSession() as session:
            async with session.post(self.webhook_url, json=payload) as response:
                if response.status != 204:
                    logging.error(f"Discord notification failed: {response.status}")

class TelegramNotifier:
    def __init__(self, config):
        self.bot_token = config['bot_token']
        self.chat_ids = config['chat_ids']
        
    async def send_alert(self, alert: Alert):
        emoji_map = {
            AlertSeverity.LOW: "🟢",
            AlertSeverity.MEDIUM: "🟡",
            AlertSeverity.HIGH: "🟠",
            AlertSeverity.CRITICAL: "🔴"
        }
        
        message = f"""
{emoji_map[alert.severity]} *{alert.severity.value.upper()} ALERT*

*Component:* {alert.component}
*Message:* {alert.message}
*Time:* {time.strftime('%H:%M:%S', time.localtime(alert.timestamp))}

`{json.dumps(alert.details, indent=2)[:500]}`

_Quantum Trading System_
        """
        
        url = f"https://api.telegram.org/bot{self.bot_token}/sendMessage"
        
        for chat_id in self.chat_ids:
            payload = {
                "chat_id": chat_id,
                "text": message,
                "parse_mode": "Markdown"
            }
            
            async with aiohttp.ClientSession() as session:
                async with session.post(url, json=payload) as response:
                    if response.status != 200:
                        logging.error(f"Telegram notification failed: {response.status}")

class SlackNotifier:
    def __init__(self, config):
        self.webhook_url = config['webhook_url']
        self.channel = config.get('channel', '#trading-alerts')
        
    async def send_alert(self, alert: Alert):
        color_map = {
            AlertSeverity.LOW: "good",
            AlertSeverity.MEDIUM: "warning",
            AlertSeverity.HIGH: "warning",
            AlertSeverity.CRITICAL: "danger"
        }
        
        payload = {
            "channel": self.channel,
            "username": "Quantum Trading Bot",
            "icon_emoji": ":robot_face:",
            "attachments": [{
                "color": color_map[alert.severity],
                "title": f"{alert.severity.value.upper()} Alert: {alert.component}",
                "text": alert.message,
                "fields": [
                    {"title": "Timestamp", "value": time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(alert.timestamp)), "short": True}
                ],
                "footer": "Quantum Trading System",
                "ts": int(alert.timestamp)
            }]
        }
        
        if alert.details:
            payload["attachments"][0]["fields"].append({
                "title": "Details",
                "value": f"```{json.dumps(alert.details, indent=2)[:500]}```",
                "short": False
            })
            
        async with aiohttp.ClientSession() as session:
            async with session.post(self.webhook_url, json=payload) as response:
                if response.status != 200:
                    logging.error(f"Slack notification failed: {response.status}")
