requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
cupy-cuda11x>=12.2.0
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=23.2.0
uvloop>=0.19.0
websockets>=11.0.0
orjson>=3.9.0
redis>=5.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
clickhouse-driver>=0.2.6
sqlalchemy>=2.0.0
alembic>=1.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
ccxt>=4.1.0
cryptography>=41.0.0
python-multipart>=0.0.6
prometheus-client>=0.19.0
structlog>=23.2.0
hvac>=1.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
celery>=5.3.0
flower>=2.0.0
grafana-api>=1.0.3
telegram-send>=0.34
slack-sdk>=3.23.0
discord.py>=2.3.0
tweepy>=4.14.0
praw>=7.7.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
ta>=0.10.0
yfinance>=0.2.0
python-binance>=1.0.0
python-okx>=1.0.0
web3>=6.11.0
eth-account>=0.9.0
brownie-eth>=1.20.0
py-solc-x>=1.12.0
slither-analyzer>=0.10.0
mythril>=0.24.0
manticore>=0.3.7
echidna>=2.2.0
selenium>=4.15.0
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
ratelimit>=2.2.0
backoff>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.0
click>=8.1.0
rich>=13.6.0
typer>=0.9.0
schedule>=1.2.0
apscheduler>=3.10.0
kubernetes>=28.1.0
docker>=6.1.0
paramiko>=3.3.0
fabric>=3.2.0
invoke>=2.2.0
dask[complete]>=2023.10.0
ray[default]>=2.8.0
mlflow>=2.7.0
wandb>=0.16.0
optuna>=3.4.0
hyperopt>=0.2.7
bayesian-optimization>=1.4.0
catboost>=1.2.0
keras>=2.14.0
tensorflow>=2.14.0
pytorch-lightning>=2.1.0
gymnasium>=0.29.0
stable-baselines3>=2.1.0
zipline-reloaded>=3.0.0
pyfolio-reloaded>=0.9.0
empyrical>=0.5.0
quantlib>=1.32.0
arch>=6.2.0
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0
sktime>=0.24.0
tsfresh>=0.20.0
pykalman>=0.9.5
filterpy>=1.4.5
ta-lib>=0.4.0
finta>=1.3.0
technical>=1.4.0
jupyter>=1.0.0
ipywidgets>=8.1.0
plotly>=5.17.0
dash>=2.14.0
streamlit>=1.28.0
bokeh>=3.3.0
seaborn>=0.13.0
matplotlib>=3.8.0
networkx>=3.2.0
igraph>=0.10.0
pyvis>=0.3.0
folium>=0.15.0
geopandas>=0.14.0
geopy>=2.4.0